Enable rollup for types
By enabling rollup for typescript types we ensure that the build declaration file is at dist/index.d.ts.
Otherwise the declaration files would be kept in the same structure as our source which could be problematic if the src structure changed, leading to a change in the resulting structure of our built files (in dist/) which could be different to the path definitions in our package.json.