fix(get): reactivity on proxy objects
Fixes #34

This is a backport of the change from the 2.x version. Thanks to @Tofandel for reporting the issue.

We simply just added a condition to the reduce code to actually touch the objects instead of only checking the path existence via the `hasOwnProperty` call.

This enables frameworks like Vue to properly track the reactivity on return values again.